Split behavior to internal and shared
